backend/requirements/dev.txt
-r base.txt
pytest>=7.3.1,<8.0.0
pytest-django>=4.5.2,<5.0.0
black>=23.3.0,<24.0.0
flake8>=6.0.0,<7.0.0
isort>=5.12.0,<6.0.0
django-debug-toolbar>=4.0.0,<5.0.0
ipython>=8.13.0,<9.0.0
